added generateMetadata function to /challengerwelcome page
